Added grouped_matrix_writer.py
I have a rectangular `.csv` table/matrix $M$ with elements $M_{ij}$:

```
M_11,M_12,M_13,...
M_21,M_22,M_23,...
.
.
.
```

where the `.`s are placeholders; please note that there is no spacing between each two adjacent elements. I need a Python 3.10+ and/or bash script that writes it to a `.txt` file like this:

```
1 => (M_11, M_12, M_13,
      M_14, M_15, M_16,
      ...)
2 => (M_21, M_22, M_23,
      M_24, M_25, M_26,
      ...)
...
```

please note that there is a space character between each two adjacent elements. So, if we have a 150x150 matrix then we would go down to `150 =>` and there would be $150 ÷ 3 = 50$ code lines for each row in the `.csv` file.